some scripts don't need to run on MDs

+++ src/PlutoUpgradeHelper/002-000-011-DisklessFsMovedToHome.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
 # Changed from 2.0.0.10 to 2.0.0.11: diskless filesystems were moved to /home
+if [[ -f /etc/diskless.conf ]] ;then
+	exit 0
+fi
+
 if [[ ! -L /usr/pluto/diskless && -d /usr/pluto/diskless ]]; then
         mv /usr/pluto/diskless /home
 else
